+++ lending/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v1.2.0] lending
 ### Releases
 - [NPM v1.2.0] https://www.npmjs.com/package/@codat/lending/v/1.2.0 - lending
+
+## 2023-09-14 15:08:43
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.83.2 (2.109.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] lending
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@codat/lending/v/2.1.0 - lending